Squashed 'software/src/master/' changes from 96a9ec0..29fc41d
29fc41d Merge branch 'master' into release-8.1 to copy new LICENSE file.
6534d12 Merge 'software/src/8.0' development into 'software/src/8.1' code base.
d470ed0 Delete 'bin' sub-directory from software/src/8.1.
131063e Merge branch 'master' (solaris-11.3) into macport
22eb2bd Merge branch 'master' of github.com:MichaelJCaruso/vision
10a5e73 Null merge of branch 'solaris-11.3-x86' into macport
88b4a5b Merge 8.1 '-U_FORTIFY_SOURCE' 'master' changes into 'solaris-11.3-x86'
ba3c43e Merge 8.1 '-U_FORTIFY_SOURCE' 'master' changes into 'macport'
eaa0212 Update 8.1 to use -U_FORTIFY_SOURCE Linux compiler option.
c462cfb .gitignore extra files generated by 'c++ -g' on the Mac.
c7c3ee7 Merge branch '8.1-bootstrap-fix' into staging
7220a0c Restore definition of 'setColumnPtr:' debug method.
5309049 Duplicate ad5c6309e763f9112e3c138beeb4b3fb101b6f4c changes under 8.1.
49d5efd Fix at-exit crashes of 8.1 executables on linux.
78ac3a0 Add output of 'visionBuilder debug' to .gitignore.
a9d6a01 Vision 8.1 for github!!

git-subtree-dir: software/src/master
git-subtree-split: 29fc41d26710a60efa1dbc630bcd96a730817b40

+++ src/backend/VOutputBuffer.h
@@ -171,7 +171,7 @@ public:
 	/*----  VOutputBuffer::AllocatoR::CellBlock::Cell  ----*/
 	//  Allocation
 	protected:
-	    static unsigned int g_iTotalAllocation;
+	    static size_t g_iTotalAllocation;
 	public:
 	    void* operator new (size_t sStructure, size_t sCellBlock) {
 		g_iTotalAllocation += sCellBlock;
